Move AppBookCard component to matters dir
